+++ test/eca/features/chat_test.clj
@@ -230,7 +230,7 @@
                     f.prompt/get-prompt! (fn [_ args-map _]
                                            (reset! prompt-args args-map)
                                            {:messages [{:role :user :content "test"}]})
-                    f.chat/prompt-messages! (fn [messages _ ctx] (reset! invoked? [messages ctx]))]
+                    f.chat/prompt-messages! (fn [messages ctx] (reset! invoked? [messages ctx]))]
         (#'f.chat/send-mcp-prompt! {:prompt "awesome-prompt" :args [42 "yo"]} test-chat-ctx)
         (is (match?
              @prompt-args